requirements.txt
appdirs
asgiref
astroid
autopep8
certifi
chardet
colorama
cvlib
cycler
decorator
distlib
dj-database-url
django-crispy-forms
django-debug-toolbar
django-filter
django-heroku
easy-thumbnails
face-recognition-models
filelock
gunicorn
idna
imageio
imutils
isort
joblib
kiwisolver
lazy-object-proxy
matplotlib
mccabe
mysql-connector-python
networkx
numpy
opencv-contrib-python
opencv-python
pandas
Pillow
pipenv
progressbar
protobuf
psycopg2
pycodestyle
pylint
pyparsing
python-dateutil
python-mnist
pytz
PyWavelets
requests
scikit-image
scikit-learn
scipy
six
sqlparse
threadpoolctl
tifffile
toml
urllib3
virtualenv
virtualenv-clone
whitenoise
wincertstore
wrapt
